Insert Bug solucionado * Solucion al problema del render en el agregado del Bus * Creacion del Custom Select

+++ src/context/busContext.js
@@ -1,9 +1,11 @@
 import React, { createContext, useContext, useEffect, useState } from "react";
 import { firestore } from "firebase";
-import { SeatDesignContext } from "./seatDesignContext";
+import useRemoveThis from "../hooks/useRemoveThis";
+
 export const BusContext = createContext();
 export const BusProvider = (props) => {
   const [bus, setBus] = useState([]);
+  const question = useRemoveThis();
 
   const requestUnsubscribe = snapshot => {
     const listBus = snapshot.docs.map((doc) => ({
@@ -18,13 +20,34 @@ export const BusProvider = (props) => {
     return () => unsubscribe();
   }, []);
 
-  const deleteBus = (id) => {
-    firestore().collection("bus").doc(id).delete();
+  const updateActive = async (id,state = "Si") => {
+    const ref = firestore().collection('bus');
+
+    console.log(state);
+
+    try{
+      switch(state) {
+        case "Si":
+          await ref.doc(id).update({ active : false });
+          break;
+        case "No":
+          await ref.doc(id).update({ active : true });
+          break;
+      }
+    }catch(e){
+      console.log(e);
+    }
+  }
+
+  const deleteBus = async (id) => {
+    const security = await question();
+    if(security){
+      await firestore().collection("bus").doc(id).delete();
+    }
   };
 
-
   return (
-    <BusContext.Provider value={{ bus, deleteBus }}>
+    <BusContext.Provider value={{ bus, deleteBus , updateActive }}>
       {props.children}
     </BusContext.Provider>
   );
